streamlit_app/requirements.txt
streamlit>=1.22.0
google-cloud-storage>=2.8.0
requests>=2.28.2
python-dotenv>=1.0.0
